html/semantics/embedded-content/the-iframe-element/iframe-loading-lazy-nav-link-click-fragment.html

<!DOCTYPE html>
<title>Navigating iframe loading='lazy' before it is loaded: link click (fragment)</title>
<iframe name="iframe" src="support/blank.htm?src" loading="lazy"></iframe>
<a href="about:blank#fragment" target="iframe"></a>
<script>
const iframe = document.querySelector('iframe');
const a = document.querySelector('a');
a.click();
</script>
<!-- Loading testharness.js here is intentional. -->
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
setup({single_test: true});
onload = () => {
  step_timeout(() => {
    assert_equals(iframe.contentWindow.location.href, "about:blank#fragment");
    done();
  }, 250);
}
</script>
<!-- delay the load event -->
<iframe src="support/blank.htm?src"></iframe>
