LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-iframe-element/iframe-loading-lazy-nav-link-click-fragment-expected.txt



FAIL Navigating iframe loading='lazy' before it is loaded: link click (fragment) Error: assert_equals: expected "about:blank#fragment" but got "http://web-platform.test:8800/html/semantics/embedded-content/the-iframe-element/support/blank.htm?src"
